Pull tout fonctionne avant qu'on casse le code

+++ VAO.h
@@ -81,7 +81,7 @@ public:
 		return objects.at(index);
 	}
 
-	inline unsigned getMeshAmount() {
+	inline size_t getMeshAmount() {
 		return objects.size();
 	}
 };
